refactor(binding-kafka): rename AlterConfigs/DescribeConfigs api classes with Kafka prefix
AlterConfigsResponse(V2FW) and DescribeConfigsResponse(V4FW), added by this
PR, become KafkaAlterConfigsResponse(V2FW) and KafkaDescribeConfigsResponse(V4FW)
to match the Kafka* naming already used by every other class in the
binding-kafka api package. Pre-existing CreateTopicsResponse(V7FW) and
DeleteTopicsResponse(V6FW), which predate this PR and also lack the prefix,
are left as-is.
