+++ runtime/binding-kafka/src/main/java/io/aklivity/zilla/runtime/binding/kafka/api/KafkaDescribeConfigsResponse.java
@@ -28,7 +28,7 @@ import io.aklivity.zilla.runtime.common.agrona.buffer.DirectBufferEx;
  * boundary.
  * </p>
  */
-public interface DescribeConfigsResponse
+public interface KafkaDescribeConfigsResponse
 {
     enum Kind
     {
